% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticdump.R
\name{elasticdump}
\alias{elasticdump}
\title{Perform a basic elasticdump call}
\usage{
elasticdump(src, dest, limit = 100, type = "data", rm_dest = TRUE)
}
\arguments{
\item{src}{The data source; may be http or file (see Details)}

\item{dest}{The data destination; may be http or file (see Details)}

\item{limit}{The max. number of docs to dump per batch}

\item{type}{One of data (default), mapping, or analyzer}
}
\value{
The return code from `elasticdump`; 0 = success
}
\description{
Perform a basic elasticdump call
}
\details{
Elasticdump provides basic transfer of elasticsearch indices/types
to and from http and file. The \code{src} may be an elasticsearch server
}

