% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticdump.R
\name{es_data_backup}
\alias{es_data_backup}
\title{Back up the data from an elasticsearch index + type}
\usage{
es_data_backup(type, server, index, bak_dir)
}
\arguments{
\item{type}{Elasticsearch type to retrieve}

\item{server}{Http/s elasticsearch server address}

\item{index}{Name of the elasticsearch index to retrieve}

\item{bak_dir}{Path of directory to which the backup is written}
}
\value{
The return code from `elasticdump`; 0 == success
}
\description{
Back up the data from an elasticsearch index + type
}

