% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticdump.R
\name{es_mapping_backup}
\alias{es_mapping_backup}
\title{Back up the mapper of an elasticsearch index + type}
\usage{
es_mapping_backup(server, index, type, bak_dir)
}
\arguments{
\item{server}{Http/s elasticsearch server address}

\item{index}{Name of the elasticsearch index to retrieve}

\item{type}{Elasticsearch type to retrieve}

\item{bak_dir}{Path of directory to which the backup is written}
}
\value{
The return code from `elasticdump` (0 == success) or FALSE if
  \code{server} isn't an http/s address.
}
\description{
Back up the mapper of an elasticsearch index + type
}

